Rename project to fly_journey and update identifiers
