+++ android/app/build.gradle
@@ -13,7 +13,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    namespace = "com.mycompany.CounterApp"
+    namespace = "com.flyjourney.app"
     compileSdk = flutter.compileSdkVersion
     // Pin NDK to a stable installed version to avoid SDK mismatches
     ndkVersion = "26.1.10909125"
@@ -38,7 +38,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId = "com.mycompany.CounterApp"
+        applicationId = "com.flyjourney.app"
         // You can update the following values to match your application needs.
         // For more information, see: https://flutter.dev/to/review-gradle-config.
         minSdk = flutter.minSdkVersion
