--- android/app/src/main/kotlin/com/example/yap/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.yap
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()
