+++ android/app/src/main/kotlin/com/flyjourney/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.mycompany.CounterApp
+package com.flyjourney.app
 
 import io.flutter.embedding.android.FlutterActivity
