+++ lib/src/core/widgets/flight_card.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:cnh_n/src/features/search/domain/models/flight.dart';
+import 'package:fly_journey/src/features/search/domain/models/flight.dart';
 
 class FlightCard extends StatelessWidget {
   final Flight flight;
