+++ lib/src/features/booking/presentation/screens/booking_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:cnh_n/src/features/search/domain/models/flight.dart';
-import 'package:cnh_n/src/features/booking/domain/models/passenger.dart';
-import 'package:cnh_n/src/features/booking/domain/models/booking.dart';
-import 'package:cnh_n/src/features/booking/data/services/storage_service.dart';
+import 'package:fly_journey/src/features/search/domain/models/flight.dart';
+import 'package:fly_journey/src/features/booking/domain/models/passenger.dart';
+import 'package:fly_journey/src/features/booking/domain/models/booking.dart';
+import 'package:fly_journey/src/features/booking/data/services/storage_service.dart';
 
 class BookingScreen extends StatefulWidget {
   final Flight flight;
