+++ lib/src/features/booking/presentation/screens/passenger_information_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:cnh_n/src/features/search/domain/models/flight.dart';
-import 'package:cnh_n/src/core/constants/colors.dart';
-import 'package:cnh_n/src/features/booking/presentation/screens/payment_screen.dart';
-import 'package:cnh_n/src/features/booking/domain/models/passenger_models.dart';
+import 'package:fly_journey/src/features/search/domain/models/flight.dart';
+import 'package:fly_journey/src/core/constants/colors.dart';
+import 'package:fly_journey/src/features/booking/presentation/screens/payment_screen.dart';
+import 'package:fly_journey/src/features/booking/domain/models/passenger_models.dart';
 
 class PassengerInformationScreen extends StatefulWidget {
   final Flight outboundFlight;
