+++ lib/src/features/booking/presentation/screens/payment_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:cnh_n/src/features/search/domain/models/flight.dart';
-import 'package:cnh_n/src/core/constants/colors.dart';
-import 'package:cnh_n/src/features/booking/domain/models/passenger_models.dart';
-import 'package:cnh_n/src/features/booking/domain/models/booking.dart';
-import 'package:cnh_n/src/features/booking/domain/models/passenger.dart';
-import 'package:cnh_n/src/features/booking/data/services/storage_service.dart';
-import 'package:cnh_n/src/features/booking/presentation/screens/payment_success_screen.dart';
+import 'package:fly_journey/src/features/search/domain/models/flight.dart';
+import 'package:fly_journey/src/core/constants/colors.dart';
+import 'package:fly_journey/src/features/booking/domain/models/passenger_models.dart';
+import 'package:fly_journey/src/features/booking/domain/models/booking.dart';
+import 'package:fly_journey/src/features/booking/domain/models/passenger.dart';
+import 'package:fly_journey/src/features/booking/data/services/storage_service.dart';
+import 'package:fly_journey/src/features/booking/presentation/screens/payment_success_screen.dart';
 
 class PaymentScreen extends StatefulWidget {
   final Flight outboundFlight;
