+++ lib/src/features/search/presentation/screens/search_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:cnh_n/src/core/widgets/header.dart';
-import 'package:cnh_n/src/core/widgets/footer.dart';
+import 'package:fly_journey/src/core/widgets/header.dart';
+import 'package:fly_journey/src/core/widgets/footer.dart';
 
 class SearchScreen extends StatelessWidget {
   final String? from;
