+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: cnh_n
+name: fly_journey
 description: "Khám phá thế giới với ứng dụng đặt vé máy bay tiện lợi của chúng tôi, giúp bạn tìm kiếm, so sánh và đặt vé dễ dàng cho mọi hành trình."
 publish_to: "none"
 version: 1.0.0
